Add files via upload
The logins.csv is the database. I used a login object that consists of a username and password. In the CSVReadWrite file there are three methods. The first is the write method which is essentially the sign up method that adds the username and password to the csv file. The read file just displays all the usernames and passwords and the check method is the one that checks if the user and pass you put in are actually in the database and correspond with a login object.